facts: server.Mounts: fix whitespaces and escaped characters
The `mount` command can not be parsed unambigiously. Whitespaces in paths
and , in option lists are not escaped. See the new tests for examples
the old code can not handle.

This patch switches to reading from `/proc/self/mountinfo` instead (see man 5
proc for documentation) where strings are properly escaped.

While `/proc/self/mountinfo` provides a lot more details, these are
ignored to keep compatibility with the old type. If there is a need, the
missing fields can be easily added.