add metadata spec for pyinfra plugins
This is a start on improving plugin discovery in pyinfra. So far operations/facts have all been builtins and the lack of discovery/organisation has meant little sharing of external ops/facts (ie, plugins). As an example the metadata structure here might provide a way to include plugins within the main pyinfra documentation. There's a lot of potential future improvements in this area.

This also hugely improves the docs by adding some categorisation to ops/facts with filtering UX.

Co-authored-by: Nick Mills-Barrett <nick@fizzadar.com>